+++ data/priv/repo/migrations/20250328060100_create_contributor_xp.exs
@@ -4,12 +4,12 @@ defmodule Moc.Data.Repo.Migrations.CreateContributorXp do
   def change do
     create table("contributor_xp") do
       add :xp, :float, null: false
-      add :charisma, :integer, null: false, default: 0
-      add :constitution, :integer, null: false, default: 0
+      add :charisma, :float, null: false, default: 0
+      add :constitution, :float, null: false, default: 0
       add :dark_xp, :float, null: false
-      add :dexterity, :integer, null: false, default: 0
+      add :dexterity, :float, null: false, default: 0
       add :light_xp, :float, null: false
-      add :wisdom, :integer, null: false, default: 0
+      add :wisdom, :float, null: false, default: 0
       add :pull_request_id, references("pull_requests", on_delete: :delete_all), null: false
       add :contributor_id, references("contributors", on_delete: :delete_all), null: false
